Made the python and matlab code consistent
